+++ src/main/kotlin/dev/aaronhowser/mods/geneticsresequenced/api/genes/GeneRegistry.kt
@@ -18,4 +18,22 @@ object GeneRegistry {
     fun fromId(id: ResourceLocation): Gene? = GENE_REGISTRY.get(id)
     fun fromId(id: String): Gene? = GENE_REGISTRY.get(ResourceLocation.parse(id))
 
+    fun fromIdPath(path: String): Gene? = GENE_REGISTRY.find { it.id.path == path }
+
+    fun getRegistrySorted(): List<Gene> {
+        val mutations = mutableListOf<Gene>()
+        val negatives = mutableListOf<Gene>()
+        val other = mutableListOf<Gene>()
+
+        for (gene in GENE_REGISTRY) {
+            when {
+                gene.isMutation -> mutations.add(gene)
+                gene.isNegative -> negatives.add(gene)
+                else -> other.add(gene)
+            }
+        }
+
+        return other.sortedBy { it.id } + mutations.sortedBy { it.id } + negatives.sortedBy { it.id }
+    }
+
 }