+++ src/main/kotlin/dev/aaronhowser/mods/geneticsresequenced/command/ListAllGenesCommand.kt
@@ -2,7 +2,6 @@ package dev.aaronhowser.mods.geneticsresequenced.command
 
 import com.mojang.brigadier.builder.ArgumentBuilder
 import com.mojang.brigadier.context.CommandContext
-import dev.aaronhowser.mods.geneticsresequenced.api.genes.Gene
 import dev.aaronhowser.mods.geneticsresequenced.api.genes.GeneRegistry
 import dev.aaronhowser.mods.geneticsresequenced.datagen.ModLanguageProvider
 import dev.aaronhowser.mods.geneticsresequenced.datagen.ModLanguageProvider.Companion.toComponent
@@ -22,7 +21,7 @@ object ListAllGenesCommand {
 
         val messageComponent = ModLanguageProvider.Commands.LIST_ALL_GENES.toComponent()
 
-        for (gene in GeneRegistry.GENE_REGISTRY) {
+        for (gene in GeneRegistry.getRegistrySorted()) {
             val geneComponent = Component
                 .literal("\n• ")
                 .append(gene.nameComponent)